submit form by enter click, refactor js code
