+++ client/index.html
@@ -53,8 +53,9 @@
   <article class="Article Article--right">
     <h2>Display compatible browsers</h2>
     <form id="query_form">
-      <textarea name="query" cols="30" rows="5" class="QueryTextArea"
-              placeholder="Write the Browserslist config here...">defaults</textarea>
+      <textarea name="query" cols="30" rows="5"
+                class="QueryTextArea" id="query_text_area"
+                placeholder="Write the Browserslist config here...">defaults</textarea>
       <button type="submit">SHOW STATS!!!</button>
     </form>
     <h2>Global coverage: <span id="global-coverage"></span></h2>
